fix taproot structure and P2WsH bug
