+++ src/transaction/random/mod.rs
@@ -5,4 +5,3 @@ pub mod script;
 pub mod transaction;
 pub mod version;
 pub mod witness;
-pub mod taproot;
